requirements.txt
numpy>=1.15.2
pandas>=0.23.4
python-twitter>=3.5
tweepy>=3.6.0
matplotlib>=3.0.0